comit

+++ Salvo/Salvo/Repositories/IGamePlayerRepository.cs
@@ -10,7 +10,7 @@ namespace Salvo.Repositories
     {
         GamePlayer GetGamePlayerView(int idGamePlayer);
         void Save(GamePlayer gamePlayer);
-        GamePlayer FindById(int id);
+        GamePlayer FindById(long id);
     }
 
 }
